VITIS-11100 Remove AIE2 logic from AIE parsing. Remove extraneous definitions from AIE2 parser.
